+++ app/lib/app.dart
@@ -1,12 +1,9 @@
-import 'package:flutter/foundation.dart' show kIsWeb;
 import 'package:flutter/material.dart';
 import 'package:flutter_gen/gen_l10n/app_localizations.dart';
-import 'package:flutter_native_timezone/flutter_native_timezone.dart';
 import 'package:provider/provider.dart';
 import 'package:studyu_flutter_common/studyu_flutter_common.dart';
-import 'package:timezone/data/latest.dart' as tz;
-import 'package:timezone/timezone.dart' as tz;
 
+import 'main.dart';
 import 'models/app_state.dart';
 import 'routes.dart';
 import 'theme.dart';
@@ -17,19 +14,10 @@ class MyApp extends StatefulWidget {
 }
 
 class _MyAppState extends State<MyApp> {
+
   @override
   void initState() {
     super.initState();
-    if (!kIsWeb) {
-      _configureLocalTimeZone();
-    }
-  }
-
-  /// This is needed for flutter_local_notifications
-  Future<void> _configureLocalTimeZone() async {
-    tz.initializeTimeZones();
-    final currentTimeZone = await FlutterNativeTimezone.getLocalTimezone();
-    tz.setLocalLocation(tz.getLocation(currentTimeZone));
   }
 
   @override
@@ -37,7 +25,7 @@ class _MyAppState extends State<MyApp> {
     return MultiProvider(
       providers: [
         ChangeNotifierProvider<AppLanguage>(create: (context) => AppLanguage(AppLocalizations.supportedLocales)),
-        Provider<AppState>(create: (context) => AppState(context)),
+        Provider<AppState>(create: (context) => AppState()),
       ],
       child: Consumer<AppLanguage>(
         builder: (context, model, child) {
@@ -50,6 +38,7 @@ class _MyAppState extends State<MyApp> {
             locale: model.appLocal,
             supportedLocales: AppLocalizations.supportedLocales,
             localizationsDelegates: AppLocalizations.localizationsDelegates,
+            navigatorKey: navigatorKey,
           );
         },
       ),