+++ app/lib/screens/study/dashboard/dashboard.dart
@@ -5,7 +5,6 @@ import 'package:flutter_gen/gen_l10n/app_localizations.dart';
 import 'package:material_design_icons_flutter/material_design_icons_flutter.dart';
 import 'package:package_info_plus/package_info_plus.dart';
 import 'package:provider/provider.dart';
-import 'package:quiver/collection.dart';
 import 'package:studyu_core/core.dart';
 import 'package:url_launcher/url_launcher.dart';
 
@@ -15,6 +14,12 @@ import '../report/report_details.dart';
 import 'task_overview_tab/task_overview.dart';
 
 class DashboardScreen extends StatefulWidget {
+  final String error;
+
+  const DashboardScreen({
+    this.error,
+  });
+
   @override
   _DashboardScreenState createState() => _DashboardScreenState();
 }
@@ -30,17 +35,27 @@ class OverflowMenuItem {
 
 class _DashboardScreenState extends State<DashboardScreen> {
   StudySubject subject;
-  Multimap<CompletionPeriod, Task> scheduleToday;
+  List<TimedTask> scheduleToday;
 
   @override
   void initState() {
     super.initState();
     subject = context.read<AppState>().activeSubject;
+    // if (subject != null) {
     scheduleToday = subject.scheduleFor(DateTime.now());
+    // }
+    if (widget.error != null) {
+      WidgetsBinding.instance.addPostFrameCallback((_) {
+        ScaffoldMessenger.of(context).showSnackBar(SnackBar(content: Text(widget.error)));
+      });
+    }
   }
 
   @override
   Widget build(BuildContext context) {
+    /* if (subject == null) {
+      return const Scaffold();
+    } */
     return Scaffold(
       appBar: AppBar(
         // Removes back button. We currently keep navigation stack to make developing easier