+++ core/lib/src/models/tasks/schedule.dart
@@ -1,4 +1,5 @@
 import 'package:json_annotation/json_annotation.dart';
+import 'package:uuid/uuid.dart';
 
 part 'schedule.g.dart';
 
@@ -7,7 +8,7 @@ typedef ScheduleParser = Schedule Function(Map<String, dynamic> data);
 @JsonSerializable()
 class Schedule {
   List<CompletionPeriod> completionPeriods = [
-    CompletionPeriod(
+    CompletionPeriod.withId(
       unlockTime: StudyUTimeOfDay(hour: 8),
       lockTime: StudyUTimeOfDay(hour: 20),
     )
@@ -28,17 +29,20 @@ class Schedule {
 
 @JsonSerializable()
 class CompletionPeriod {
+  // Todo make non-nullable (breaks backwards compatibility)
+  String? id;
   final StudyUTimeOfDay unlockTime;
   final StudyUTimeOfDay lockTime;
 
   CompletionPeriod({required this.unlockTime, required this.lockTime});
+  CompletionPeriod.withId({required this.unlockTime, required this.lockTime}) : id = const Uuid().v4();
 
   factory CompletionPeriod.fromJson(Map<String, dynamic> json) => _$CompletionPeriodFromJson(json);
 
   Map<String, dynamic> toJson() => _$CompletionPeriodToJson(this);
 
   @override
-  String toString() => '${unlockTime.toString()} - ${lockTime.toString()}';
+  String toString() => '$unlockTime - $lockTime';
 
   bool contains(StudyUTimeOfDay time) {
     return unlockTime.earlierThan(time) && time.earlierThan(lockTime);
@@ -50,6 +54,17 @@ class StudyUTimeOfDay {
   int minute = 0;
 
   StudyUTimeOfDay({this.hour = 0, this.minute = 0}) : super();
+
+  StudyUTimeOfDay.fromDateTime(DateTime date) {
+    hour = date.toLocal().hour;
+    minute = date.toLocal().minute;
+  }
+
+  StudyUTimeOfDay.now() {
+    final now = StudyUTimeOfDay.fromDateTime(DateTime.now());
+    hour = now.hour;
+    minute = now.minute;
+  }
 
   StudyUTimeOfDay.fromJson(String value) {
     final elements = value.split(':').map(int.parse);