+++ app/CHANGELOG.md
@@ -1,3 +1,24 @@
+## 1.11.0
+
+ - **FIX**: show correct screen when app is launched by a notification. ([77c22707](https://github.com/hpi-studyu/studyu/commit/77c227078fa70e78e29ee7bf810b2bc20cb45657))
+ - **FIX**: do not show interventions with empty titles. ([2a9abe76](https://github.com/hpi-studyu/studyu/commit/2a9abe764d245062548cff5a1a43251a9fc14a4f))
+ - **FIX**: tasks cannot be completed multiple times via notifications + localization. ([d9974911](https://github.com/hpi-studyu/studyu/commit/d997491127016fed840a6b9b336db9b98583c051))
+ - **FIX**: restructure notifications. ([e6627f5a](https://github.com/hpi-studyu/studyu/commit/e6627f5a4ae635465a1b805338b8b2f122704d03))
+ - **FIX**: report screen null exceptions. ([38c59426](https://github.com/hpi-studyu/studyu/commit/38c5942626d042d6287fec8c5d603999c421193a))
+ - **FIX**: keep alive question item states on scrolling. ([c27881f8](https://github.com/hpi-studyu/studyu/commit/c27881f8102f27a3200fc3e3ad7828ef1061edda))
+ - **FIX**: tasks could be accessed outside of their completion times. ([d6fcd920](https://github.com/hpi-studyu/studyu/commit/d6fcd920af2f14ed57235111b7382f736cffbb55))
+ - **FIX**: task screen empty title null exception. ([f8527dd4](https://github.com/hpi-studyu/studyu/commit/f8527dd41f16ec16ec7d8c5a8f36eb7e15775f0b))
+ - **FIX**: notification icon. ([523e6396](https://github.com/hpi-studyu/studyu/commit/523e6396d143962e50eaaccc91fd2fc0cc7adb44))
+ - **FIX**: notification not being delivered, add notification body. ([ed8e440b](https://github.com/hpi-studyu/studyu/commit/ed8e440ba66e04e0c06e98b4e1c0770927855377))
+ - **FIX**: Skip consentCheck if study doesn't have one. ([5675d92a](https://github.com/hpi-studyu/studyu/commit/5675d92af89849bf8fb22ee99bfcf5a61408b5a4))
+ - **FIX**: Remove error message at performance section. ([6a5b5905](https://github.com/hpi-studyu/studyu/commit/6a5b5905bc2aa765e550c37225f111603b692aa1))
+ - **FIX**: upgrade to flutter 3.7.0. ([2aa80015](https://github.com/hpi-studyu/studyu/commit/2aa80015a22161773649790edf2c38742809c45b))
+ - **FIX**: refactor questionnaire widget for eligibility and related tasks. ([b9639bcc](https://github.com/hpi-studyu/studyu/commit/b9639bcce5607b975b592683dc0c5d9a552f3927))
+ - **FIX**: upgrade deps. ([1b411140](https://github.com/hpi-studyu/studyu/commit/1b41114080e405c385c15b7beb8d140193656e6a))
+ - **FIX**: lint errors. ([6007e46e](https://github.com/hpi-studyu/studyu/commit/6007e46e45c36e77932930ce94c634499d57a444))
+ - **FEAT**: Upgrade notifications & deps. ([08be0c3a](https://github.com/hpi-studyu/studyu/commit/08be0c3ac3bcffabd9ad6b518cbc0e4067183424))
+ - **FEAT**: - Results of tasks with multiple completion periods are now saved uniquely. ([bf9e2818](https://github.com/hpi-studyu/studyu/commit/bf9e2818b5dc88102f69a7e773b5f9ccea0e7c73))
+
 ## 1.10.21
 
  - **FIX**: upgrade deps. ([d221eae7](https://github.com/hpi-studyu/studyu/commit/d221eae7ea55fa8761d7ff7625f92eaa3488148e))